% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_template.R
\name{fair_template}
\alias{fair_template}
\title{FAIR JSON Templates}
\usage{
fair_template(
  df,
  dataset_name = as.character(substitute(df)),
  lookups = c(),
  title = "",
  description = "",
  creator = "",
  contactPoint = ""
)
}
\arguments{
\item{df}{A data.frame}

\item{dataset_name}{Name of the dataset to be shown in the library}

\item{lookups}{Indexes of columns to create lookups for}
}
\description{
Create FAIR JSON templates from existing data
}
