% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lacunes_infarcts.R
\name{lacunes_infarcts}
\alias{lacunes_infarcts}
\title{Lacunes and Infarctions}
\usage{
lacunes_infarcts(n, visit = "V1", missing = 0.1, ids = NULL, visit_ids = NULL)
}
\description{
Lacunes and Infarctions
}
