% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socio_demographics.R
\name{socio_demographics}
\alias{socio_demographics}
\title{Socio-Demographics}
\usage{
socio_demographics(n, ids = NULL)
}
\arguments{
\item{n}{Integer number of participants to create data for}

\item{ids}{Optional vector of participant IDs}
}
\value{
A data.frame
}
\description{
Socio-Demographics
}
