% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital_signs.R
\name{vital_signs}
\alias{vital_signs}
\title{Vital Signs}
\usage{
vital_signs(n, visit = "V1", missing = 0.1, ids = NULL, visit_ids = NULL)
}
\arguments{
\item{n}{Integer number of participants to create data for}

\item{visit}{The name of the visit to create data for as a character string}

\item{missing}{The proportion of participants for which the clinical measurements were not taken completed}

\item{ids}{Optional vector of participant IDs}

\item{visit_ids}{Optional vector of visit IDs}
}
\value{
A data.frame
}
\description{
Vital Signs
}
