% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_supermarket_trolley.R
\name{vr_supermarket_trolley}
\alias{vr_supermarket_trolley}
\title{Virtual Reality Supermarket Trolley}
\usage{
vr_supermarket_trolley(
  n,
  visit = "V1",
  missing = 0.1,
  ids = NULL,
  visit_ids = NULL
)
}
\description{
Virtual Reality Supermarket Trolley
}
