% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lacunes_infarcts.R
\name{lacunes_infarcts}
\alias{lacunes_infarcts}
\title{Lacunes and Infarctions}
\usage{
lacunes_infarcts(n, visit = "V1", missing = 0.1, ids = NULL, visit_ids = NULL)
}
\arguments{
\item{n}{Integer number of participants to create data for}

\item{visit}{The name of the visit to create data for as a character string}

\item{missing}{The proportion of participants who did not receive an MRI}

\item{ids}{Optional vector of participant IDs}

\item{visit_ids}{Optional vector of visit IDs}
}
\value{
A data.frame
}
\description{
Lacunes and Infarctions
}
